pyproject.toml
[project]
name = "ai_debate_club"
version = "0.1.0"
description = "AI Debate Club application"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Oren Segal", email = "orenssegal@gmail.com"}
]

dependencies = [
    "pydantic>=2.0.0",
    "streamlit==1.32.0",
    "langchain>=0.1.0",
    "langgraph==0.0.20",
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    "chromadb>=0.4.0",
    "polars==0.19.12",
    "wikipedia==1.4.0",
    "python-dotenv==1.0.0",
    "langchain-community>=0.0.10",
    "langchain_openai==0.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch]
version = "0.1.0"

[tool.hatch.build.targets.wheel]
only-include = ["ai_debate_club"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = 'py39'
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
show_error_codes = true
